fonts: Store `ByteIndex` as a `usize`
This value is supposed to represent an UTF-8 code point offset or
length. It should never be negative and we often need to convert it to a
`usize`, so we can store it as a `usize`.

+++ components/fonts/font.rs
@@ -481,7 +481,7 @@ impl Font {
             });
 
             let glyph = GlyphData::new(glyph_id, advance, offset, true, true);
-            glyphs.add_glyph_for_byte_index(ByteIndex(i as isize), character, &glyph);
+            glyphs.add_glyph_for_byte_index(ByteIndex(i), character, &glyph);
             prev_glyph_id = Some(glyph_id);
         }
         glyphs.finalize_changes();
